requirements.txt
flask==2.0.1
python-dotenv==0.19.0
supabase==1.0.3
mistralai==0.0.7
feedparser==6.0.10
pytz==2023.3
python-dateutil==2.8.2
deep-translator==1.11.4
langdetect==1.0.9
pyyaml==6.0.1 